aioredlock: Send the scripts when evaluating them
I propose this change in order to add reliability to aioredlock.
If redis is shipped or cleared  while a loop is locking redis, then we will go into an inconsistent state whereby aioredlock thinks the scripts exist on the server, where they actually don't.

In my opinion, storing scripts in redis for these locking operations introduces unnecessary complexity by having state in redis while it's not really needed. Also, the performance will be very similar, as the scripts are very small anyway. Other implementations of redlock like redlock-py, redlock are not storing scripts.