Here I go solving problems with recursion again, oh boy
So, all of the logic to handle the evalsha already exists in the
set_lock and unset_lock functions. If we catch a ReplyError with
NOSCRIPT, that means the EVALSHA script does not exist on the redis
server, so we will just try and reregister the script, and then execute
again.

There should not be a problem where we have a chance to hit the python
recursion limit, unless something is really wrong with the redis server.