feat(redis): Enhance exceptions usage
In PR #93, I proposed to use exceptions chaining so we can use
`exception.__cause__` in the calling code to determine which original
exception occured and act upon.

But from a user's perspective, it's not easy to use at all.
What would be better is to raise specific exceptions we distinguished
in the previous commit so we can catch them with multiple `except`
statements in our calling code.

That's what this commit implements.
I've also updated usage examples to show the simplification.